requirements.txt
django==2.1.12
djangorestframework==3.10.3
dj-database-url==0.5.0
psycopg2-binary==2.8.3

